backend/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.5.0
django-cors-headers==3.5.0
Pillow==7.2.0